python/requirements.txt
fastapi
uvicorn
apscheduler
yfinance
pytz
requests
python-dotenv
selenium
